magento/magento2-functional-testing-framework#54: Set a global value for timeouts that's used across all wait related actions
- Reverted MagentoWebDriver class
- Moved getDefaultWaitTimeout method to ActionObject class
- Changed generateStepsPhp method in TestGenerator class to use timeout from getDefaultWaitTimeout method if not set
- Moved and fixed unit test for the implementation
