+++ .env.example
@@ -36,5 +36,5 @@ MODULE_WHITELIST=Magento_Framework,Magento_ConfigurableProductWishlist,Magento_C
 #MFTF_DEBUG=
 
 #*** Default timeout for wait actions
-WAIT_TIMEOUT=20
+WAIT_TIMEOUT=10
 #*** End of .env ***#
