+++ src/Magento/FunctionalTestingFramework/Test/Objects/ActionObject.php
@@ -42,6 +42,7 @@ class ActionObject
     const ACTION_ATTRIBUTE_VARIABLE_REGEX_PARAMETER = '/\(.+\)/';
     const ACTION_ATTRIBUTE_VARIABLE_REGEX_PATTERN = '/{{[\w]+\.[\w\[\]]+}}/';
     const ACTION_ATTRIBUTE_VARIABLE_REGEX_PATTERN_WITH_PARAMS= '/{{[\w]+\.[\w]+\(.+\)}}/';
+    const DEFAULT_WAIT_TIMEOUT = 10;
 
     /**
      * The unique identifier for the action
@@ -126,6 +127,16 @@ class ActionObject
         if ($order == ActionObject::MERGE_ACTION_ORDER_AFTER) {
             $this->orderOffset = 1;
         }
+    }
+
+    /**
+     * Retrieve default timeout in seconds for 'wait*' actions
+     *
+     * @return int
+     */
+    public static function getDefaultWaitTimeout()
+    {
+        return getenv('WAIT_TIMEOUT') ?: self::DEFAULT_WAIT_TIMEOUT;
     }
 
     /**
