+++ src/Magento/FunctionalTestingFramework/Util/TestGenerator.php
@@ -492,6 +492,7 @@ class TestGenerator
             if (isset($customActionAttributes['timeout'])) {
                 $time = $customActionAttributes['timeout'];
             }
+            $time = $time ?? ActionObject::getDefaultWaitTimeout();
 
             if (isset($customActionAttributes['parameterArray']) && $actionObject->getType() != 'pressKey') {
                 // validate the param array is in the correct format
